+++ src/actions.ts
@@ -1,5 +1,5 @@
 import { ActionContext, CommitOptions, DispatchOptions } from 'vuex';
-import { ActionResults, ObservedCommit, ObservedDispatch, actionFn, commitFn, dispatchFn, ActionCtx } from './index';
+import { ActionResults, ObservedCommit, ObservedDispatch, actionFn, commitFn, dispatchFn, ActionCtx, actionPromise } from './index';
 
 export const action = <S, R>(
     action: actionFn<S, R>,
@@ -11,13 +11,30 @@ export const action = <S, R>(
     ctx.commit = emitCommit(executedCommits);
     ctx.dispatch = emitDispatch(executedDispatches);
 
-    action(ctx, payload);
-
-    return expect({
-        commits: executedCommits,
-        dispatches: executedDispatches,
-    });
+    if (isPromise(action)) {
+        return expect(new Promise((resolve) => {
+            const asyncAction = action as actionPromise<S, R>;
+            asyncAction(ctx, payload).then(() => resolve({
+                commits: executedCommits,
+                dispatches: executedDispatches,
+            }));
+        }));
+    } else {
+        action(ctx, payload);
+        
+        return expect({
+            commits: executedCommits,
+            dispatches: executedDispatches,
+        });
+    }
 };
+
+const isPromise = <T extends any>(obj: any): obj is PromiseLike<T> =>
+    !!obj && (
+        ((typeof obj === 'object' || typeof obj === 'function') 
+            && typeof obj.then === 'function') ||
+        obj[Symbol.toStringTag] === 'AsyncFunction'
+    );
 
 const emitCommit = (executedCommits: ObservedCommit[]): commitFn => {
     return (type: string, payload?: any, options?: CommitOptions) => {